+++ changelog.md
@@ -4,6 +4,9 @@
 ### refactor
 - create separated system service to handle elevated actions.
 
+###
+- fix ghost windows caused by a refactor donde on v2.0.13
+
 ## [2.0.14]
 ### hotfix
 - not creating the default (system) icon pack.